+++ src/ctype/CType.ts
@@ -10,111 +10,13 @@
  * @preferred
  */
 
-import * as jsonabc from 'jsonabc'
 import { CTypeWrapperModel } from './CTypeSchema'
-import * as CTypeUtils from './CTypeUtils'
-import ICType, { CompressedCType, CompressedCTypeSchema } from '../types/CType'
+import CTypeUtils from './CType.util'
+import ICType, { CompressedCType } from '../types/CType'
 import Identity from '../identity/Identity'
 import { getOwner, store } from './CType.chain'
 import TxStatus from '../blockchain/TxStatus'
 import IClaim from '../types/Claim'
-
-/**
- *  Compresses a [[CType]] schema for storage and/or messaging.
- *
- * @param cTypeSchema A [[CType]] schema object that will be sorted and stripped for messaging or storage.
- *
- * @returns An ordered array of a [[CType]] schema.
- */
-
-export function compressCTypeSchema(
-  cTypeSchema: ICType['schema']
-): CompressedCTypeSchema {
-  if (
-    !cTypeSchema.$id ||
-    !cTypeSchema.$schema ||
-    !cTypeSchema.properties ||
-    !cTypeSchema.type
-  ) {
-    throw new Error(
-      `Property Not Provided while building cTypeSchema: 
-      ${JSON.stringify(cTypeSchema, null, 2)}`
-    )
-  }
-  const sortedCTypeSchema = jsonabc.sortObj(cTypeSchema)
-  return [
-    sortedCTypeSchema.$id,
-    sortedCTypeSchema.$schema,
-    sortedCTypeSchema.properties,
-    sortedCTypeSchema.type,
-  ]
-}
-
-/**
- *  Decompresses a schema of a [[CType]] from storage and/or message.
- *
- * @param cTypeSchema A compressesd [[CType]] schema array that is reverted back into an object.
- *
- * @returns An object that has the same properties as a [[CType]] schema.
- */
-
-export function decompressCTypeSchema(
-  cTypeSchema: CompressedCTypeSchema
-): ICType['schema'] {
-  if (!Array.isArray(cTypeSchema) || cTypeSchema.length !== 4) {
-    throw new Error(
-      'Compressed cTypeSchema isnt an Array or has all the required data types'
-    )
-  }
-  return {
-    $id: cTypeSchema[0],
-    $schema: cTypeSchema[1],
-    properties: cTypeSchema[2],
-    type: cTypeSchema[3],
-  }
-}
-
-/**
- *  Compresses a [[CType]] for storage and/or messaging.
- *
- * @param cType A [[CType]] object that will be sorted and stripped for messaging or storage.
- *
- * @returns An ordered array of a [[CType]].
- */
-
-export function compressCType(cType: ICType): CompressedCType {
-  if (!cType.hash || !cType.owner || !cType.schema) {
-    throw new Error(
-      `Property Not Provided while building cType: ${JSON.stringify(
-        cType,
-        null,
-        2
-      )}`
-    )
-  }
-  return [cType.hash, cType.owner, compressCTypeSchema(cType.schema)]
-}
-
-/**
- *  Decompresses a [[CType]] from storage and/or message.
- *
- * @param cType A compressesd [[CType]] array that is reverted back into an object.
- *
- * @returns An object that has the same properties as a [[CType]].
- */
-
-export function decompressCType(cType: CompressedCType): ICType {
-  if (!Array.isArray(cType) || cType.length !== 3) {
-    throw new Error(
-      'Compressed cType isnt an Array or has all the required data types'
-    )
-  }
-  return {
-    hash: cType[0],
-    owner: cType[1],
-    schema: decompressCTypeSchema(cType[2]),
-  }
-}
 
 export default class CType implements ICType {
   public static fromCType(cTypeInput: ICType): CType {
@@ -164,7 +66,7 @@ export default class CType implements ICType {
    */
 
   public compress(): CompressedCType {
-    return compressCType(this)
+    return CTypeUtils.compressCType(this)
   }
 
   /**
@@ -174,7 +76,7 @@ export default class CType implements ICType {
    */
 
   public static decompress(cType: CompressedCType): CType {
-    const decompressedCType = decompressCType(cType)
+    const decompressedCType = CTypeUtils.decompressCType(cType)
     return CType.fromCType(decompressedCType)
   }
 }
