+++ src/requestforattestation/RequestForAttestation.ts
@@ -10,7 +10,6 @@
  * @module RequestForAttestation
  * @preferred
  */
-import * as jsonabc from 'jsonabc'
 import { v4 as uuid } from 'uuid'
 import {
   verify,
@@ -22,44 +21,16 @@ import {
 } from '../crypto/Crypto'
 
 import Identity from '../identity/Identity'
-import AttestedClaim, {
-  decompressAttestedClaim,
-  compressAttestedClaim,
-} from '../attestedclaim/AttestedClaim'
-import { compressClaim, decompressClaim } from '../claim/Claim'
+import AttestedClaim from '../attestedclaim/AttestedClaim'
+import RequestForAttestationUtils from './RequestForAttestation.util'
 import IRequestForAttestation, {
   Hash,
   NonceHash,
   ClaimHashTree,
-  CompressedNonceHash,
-  CompressedClaimHashTree,
   CompressedRequestForAttestation,
 } from '../types/RequestForAttestation'
 import { IDelegationBaseNode } from '../types/Delegation'
 import IClaim from '../types/Claim'
-import IAttestedClaim, { CompressedAttestedClaim } from '../types/AttestedClaim'
-
-function requestForAttestationErrorCheck(
-  requestForAttestation: IRequestForAttestation
-): void {
-  if (
-    !requestForAttestation.claim ||
-    !requestForAttestation.legitimations ||
-    !requestForAttestation.claimOwner ||
-    !requestForAttestation.claimerSignature ||
-    !requestForAttestation.claimHashTree ||
-    !requestForAttestation.cTypeHash ||
-    !requestForAttestation.rootHash
-  ) {
-    throw new Error(
-      `Property Not Provided while building RequestForAttestation: ${JSON.stringify(
-        requestForAttestation,
-        null,
-        2
-      )}`
-    )
-  }
-}
 
 function hashNonceValue(nonce: string, value: any): string {
   return hashObjectAsStr(value, nonce)
@@ -94,168 +65,6 @@ function verifyClaimerSignature(reqForAtt: RequestForAttestation): boolean {
 function getHashRoot(leaves: Uint8Array[]): Uint8Array {
   const result = u8aConcat(...leaves)
   return hash(result)
-}
-
-/**
- *  Compresses an nonce and hash from a [[ClaimHashTree]] or [[RequestForAttestation]] properties.
- *
- * @param nonceHash A hash or a hash and nonce object that will be sorted and stripped for messaging or storage.
- *
- * @returns An object compressing of a hash or a hash and nonce.
- */
-
-export function compressNonceAndHash(
-  nonceHash: NonceHash
-): CompressedNonceHash {
-  if (!nonceHash.nonce || !nonceHash.hash) {
-    throw new Error(
-      `Property Not Provided while building RequestForAttestation: ${JSON.stringify(
-        nonceHash,
-        null,
-        2
-      )}`
-    )
-  }
-  return [nonceHash.hash, nonceHash.nonce]
-}
-
-/**
- *  Decompresses an nonce and hash from a [[ClaimHashTree]] or [[RequestForAttestation]] properties.
- *
- * @param nonceHash A compressesd a hash or a hash and nonce array that is reverted back into an object.
- *
- * @returns An object compressing of a hash or a hash and nonce.
- */
-
-function decompressNonceAndHash(nonceHash: CompressedNonceHash): NonceHash {
-  if (nonceHash.length === 1) {
-    return {
-      hash: nonceHash[0],
-    }
-  }
-  return {
-    hash: nonceHash[0],
-    nonce: nonceHash[1],
-  }
-}
-
-/**
- *  Compresses a [[claimHashTree]] within a [[RequestForAttestation]] object.
- *
- * @param reqForAtt A [[claimHashTree]] object that will be sorted and stripped for messaging or storage.
- *
- * @returns An ordered array of an [[claimHashTree]].
- */
-
-export function compressClaimHashTree(
-  claimHashTree: ClaimHashTree
-): CompressedClaimHashTree {
-  const sortedClaimHashTree = jsonabc.sortObj(claimHashTree)
-  const result = {}
-
-  Object.keys(sortedClaimHashTree).forEach(entryKey => {
-    result[entryKey] = compressNonceAndHash(sortedClaimHashTree[entryKey])
-  })
-  return result
-}
-
-/**
- *  Decompresses a claim hash tree from storage and/or message.
- *
- * @param reqForAtt A compressesd claim hash tree array that is reverted back into an object.
- *
- * @returns An object that has the same properties as an claim hash tree.
- */
-
-export function decompressClaimHashTree(
-  compressedClaimHashTree: CompressedClaimHashTree
-): ClaimHashTree {
-  const result = {}
-
-  Object.keys(compressedClaimHashTree).forEach(entryKey => {
-    result[entryKey] = decompressNonceAndHash(compressedClaimHashTree[entryKey])
-  })
-  return result
-}
-
-/**
- *  Compresses [[AttestedClaim]]s which are made up from an [[Attestation]] and [[RequestForAttestation]] for storage and/or message.
- *
- * @param leg An array of [[Attestation]] and [[RequestForAttestation]] objects.
- *
- * @returns An ordered array of [[AttestedClaim]]s.
- */
-
-export function compressLegitimation(
-  leg: IAttestedClaim[]
-): CompressedAttestedClaim[] {
-  return leg.map(compressAttestedClaim)
-}
-
-/**
- *  Decompresses [[AttestedClaim]]s which are an [[Attestation]] and [[RequestForAttestation]] from storage and/or message.
- *
- * @param leg A compressesd [[Attestation]] and [[RequestForAttestation]] array that is reverted back into an object.
- *
- * @returns An object that has the same properties as an [[AttestedClaim]].
- */
-
-function decompressLegitimation(
-  leg: CompressedAttestedClaim[]
-): IAttestedClaim[] {
-  return leg.map(decompressAttestedClaim)
-}
-
-/**
- *  Compresses a [[RequestForAttestation]] for storage and/or messaging.
- *
- * @param reqForAtt A [[RequestForAttestation]] object that will be sorted and stripped for messaging or storage.
- *
- * @returns An ordered array of a [[RequestForAttestation]].
- */
-
-export function compressRequestForAttestation(
-  reqForAtt: IRequestForAttestation
-): CompressedRequestForAttestation {
-  requestForAttestationErrorCheck(reqForAtt)
-  return [
-    compressClaim(reqForAtt.claim),
-    compressClaimHashTree(reqForAtt.claimHashTree),
-    compressNonceAndHash(reqForAtt.claimOwner),
-    reqForAtt.claimerSignature,
-    compressNonceAndHash(reqForAtt.cTypeHash),
-    reqForAtt.rootHash,
-    compressLegitimation(reqForAtt.legitimations),
-    reqForAtt.delegationId,
-  ]
-}
-
-/**
- *  Decompresses a [[RequestForAttestation]] from storage and/or message.
- *
- * @param reqForAtt A compressesd [[RequestForAttestation]] array that is reverted back into an object.
- *
- * @returns An object that has the same properties as a [[RequestForAttestation]].
- */
-
-export function decompressRequestForAttestation(
-  reqForAtt: CompressedRequestForAttestation
-): IRequestForAttestation {
-  if (!Array.isArray(reqForAtt) || reqForAtt.length !== 8) {
-    throw new Error(
-      'Compressed Request For Attestation isnt an Array or has all the required data types'
-    )
-  }
-  return {
-    claim: decompressClaim(reqForAtt[0]),
-    claimHashTree: decompressClaimHashTree(reqForAtt[1]),
-    claimOwner: decompressNonceAndHash(reqForAtt[2]),
-    claimerSignature: reqForAtt[3],
-    cTypeHash: decompressNonceAndHash(reqForAtt[4]),
-    rootHash: reqForAtt[5],
-    legitimations: decompressLegitimation(reqForAtt[6]),
-    delegationId: reqForAtt[7],
-  }
 }
 
 export default class RequestForAttestation implements IRequestForAttestation {
@@ -351,7 +160,9 @@ export default class RequestForAttestation implements IRequestForAttestation {
    * ```
    */
   public constructor(requestForAttestationInput: IRequestForAttestation) {
-    requestForAttestationErrorCheck(requestForAttestationInput)
+    RequestForAttestationUtils.requestForAttestationErrorCheck(
+      requestForAttestationInput
+    )
     this.claim = requestForAttestationInput.claim
     this.claimOwner = requestForAttestationInput.claimOwner
     this.cTypeHash = requestForAttestationInput.cTypeHash
@@ -558,7 +369,7 @@ export default class RequestForAttestation implements IRequestForAttestation {
    */
 
   public compress(): CompressedRequestForAttestation {
-    return compressRequestForAttestation(this)
+    return RequestForAttestationUtils.compressRequestForAttestation(this)
   }
 
   /**
@@ -570,7 +381,7 @@ export default class RequestForAttestation implements IRequestForAttestation {
   public static decompress(
     reqForAtt: CompressedRequestForAttestation
   ): RequestForAttestation {
-    const decompressedRequestForAttestation = decompressRequestForAttestation(
+    const decompressedRequestForAttestation = RequestForAttestationUtils.decompressRequestForAttestation(
       reqForAtt
     )
     return RequestForAttestation.fromRequest(decompressedRequestForAttestation)
